+++ ComicLibrary/Model/Config/Settings.cs
@@ -26,6 +26,8 @@ namespace ComicLibrary.Model.Config
 
     public string LibrariesPath { get; set; }
 
+    public bool CopyDataFromSelectedComic { get; set; } = true;
+
     public string LibrariesFilePath => Path.Combine(LibrariesPath, _librariesFileName);
 
     public string GlobalsFilePath => Path.Combine(LibrariesPath, _globalsFileName);